db/events: moved to events package
