+++ events/messages.go
@@ -1,10 +1,9 @@
-package db
+package events
 
 import (
 	"container/list"
 	"sync"
 
-	"github.com/Rugvip/bullettime/interfaces"
 	"github.com/Rugvip/bullettime/types"
 )
 
@@ -13,42 +12,47 @@ type listItem struct {
 	index uint64
 }
 
-type eventQueue struct {
-	sync.RWMutex
+type messageSource struct {
+	lock     sync.RWMutex
 	list     *list.List
 	elements map[types.Id]*list.Element
+	indices  []*list.Element
 	max      uint64
 }
 
-func NewEventStream() (interfaces.EventStream, error) {
-	return &eventQueue{
+func NewMessageSource() (*messageSource, error) {
+	return &messageSource{
 		list:     list.New(),
 		elements: map[types.Id]*list.Element{},
-		max:      1,
+		indices:  []*list.Element{},
+		max:      0,
 	}, nil
 }
 
-func (h *eventQueue) Push(event types.Event) (uint64, types.Error) {
-	h.Lock()
-	defer h.Unlock()
+func (h *messageSource) Send(event *types.Message) (uint64, types.Error) {
+	h.lock.Lock()
+	defer h.lock.Unlock()
 
 	index := h.max
 	h.max += 1
 	item := listItem{event, index}
 
-	if element := h.elements[event.Id()]; element != nil {
+	element := h.elements[event.Id()]
+	if element != nil {
+		h.indices[element.Value.(listItem).index] = element.Next()
 		h.list.MoveToFront(element)
 		element.Value = item
 	} else {
 		element = h.list.PushFront(listItem{event, h.max})
 		h.elements[event.Id()] = element
 	}
+	h.indices = append(h.indices, element)
 	return index, nil
 }
 
-func (h *eventQueue) Iterate(index uint64) ([]types.Event, types.Error) {
-	h.RLock()
-	defer h.RUnlock()
+func (h *messageSource) Iterate(index uint64) ([]types.Event, types.Error) {
+	h.lock.RLock()
+	defer h.lock.RUnlock()
 	result := []types.Event{}
 	e := h.list.Front()
 	for e != nil && e.Value.(listItem).index > index {
@@ -58,8 +62,8 @@ func (h *eventQueue) Iterate(index uint64) ([]types.Event, types.Error) {
 	return result, nil
 }
 
-func (h *eventQueue) Max() (index uint64, err types.Error) {
-	h.RLock()
-	defer h.RUnlock()
+func (h *messageSource) Max() (index uint64, err types.Error) {
+	h.lock.RLock()
+	defer h.lock.RUnlock()
 	return h.max, nil
 }
