+++ events/messages_test.go
@@ -1,4 +1,4 @@
-package db
+package events
 
 import (
 	"testing"
@@ -61,16 +61,4 @@ func (es EventStreamTest) check(from uint64, expect ...string) {
 			es.t.Error("result", i, "should be", expect[i], "was", id)
 		}
 	}
-}
-
-func typing(id string, ids ...string) types.Event {
-	event := &types.TypingEvent{}
-	event.EventType = "m.typing"
-	event.RoomId = types.NewRoomId(id, "test")
-	userIds := make([]types.UserId, len(ids))
-	for i := range ids {
-		userIds[i] = types.NewUserId(ids[i], "test")
-	}
-	event.Content.UserIds = userIds
-	return event
 }
